# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "System Security Protocols",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/protocols/msis/init": {
      "post": {
        "tags": [
          "Modified Schnor Identification Scheme"
        ],
        "operationId": "msisProtocolInit",
        "description": "Used to send A, X and obtain c",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "protocol_name": {
                    "description": "name of protocol, must be `msis`",
                    "type": "string",
                    "example": "msis"
                  },
                  "payload": {
                    "type": "object",
                    "properties": {
                      "A": {
                        "description": "stringified integer value, multiple values should be separated by single space",
                        "type": "string",
                        "example": "12345 67890"
                      },
                      "X": {
                        "description": "stringified integer value, multiple values should be separated by single space",
                        "type": "string",
                        "example": "12345 67890"
                      }
                    },
                    "required": [
                      "A",
                      "X"
                    ]
                  }
                },
                "required": [
                  "protocol_name",
                  "payload"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for `POST /protocols/msis/init` request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_token": {
                      "description": "session token generated by server",
                      "type": "string"
                    },
                    "payload": {
                      "type": "object",
                      "properties": {
                        "c": {
                          "description": "stringified integer value, multiple values should be separated by single space",
                          "type": "string",
                          "example": "12345 67890"
                        }
                      },
                      "required": [
                        "c"
                      ]
                    }
                  },
                  "required": [
                    "session_token",
                    "payload"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/protocols/msis/verify": {
      "post": {
        "tags": [
          "Modified Schnor Identification Scheme"
        ],
        "operationId": "msisProtocolCommit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "protocol_name": {
                    "description": "name of protocol, must be `msis`",
                    "type": "string",
                    "example": "msis"
                  },
                  "session_token": {
                    "description": "session token generated by server",
                    "type": "string"
                  },
                  "payload": {
                    "type": "object",
                    "properties": {
                      "S": {
                        "description": "stringified integer value, multiple values should be separated by single space",
                        "type": "string",
                        "example": "12345 67890"
                      }
                    },
                    "required": [
                      "S"
                    ]
                  }
                },
                "required": [
                  "protocol_name",
                  "session_token",
                  "payload"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response when verification was successfull.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "required": [
                    "verified"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Response when verification failed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean",
                      "example": false
                    }
                  },
                  "required": [
                    "verified"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}